pyproject.toml
[tool.poetry]
name = "agora-server"
version = "0.1.0"
description = ""
authors = ["Null Flancian <0@flancia.org"]

[tool.poetry.dependencies]
python = "^3.9"
black = "^23.3.0"
bleach = "^6.0.0"
cachelib = "^0.10.2"
cachetools = "^5.3.1"
cachetools_ext = "0.0.8"
certifi = "^2023.5.7"
charset-normalizer = "^3.1.0"
click = "^8.1.3"
dateparser = "^1.1.8"
feedgen = "^0.9.0"
feedparser = "^6.0.10"
filetype = "^1.2.0"
fuzzywuzzy = "^0.18.0"
greenlet = "^2.0.2"
idna = "^3.4"
importlib-metadata = "^6.7.0"
isodate = "^0.6.1"
itsdangerous = "^2.1.2"
jedi = "^0.18.2"
jsons = "^1.6.3"
lxml = "^4.9.2"
marko = "^2.0.0"
mdx-truly-sane-lists = "^1.3"
mypy-extensions = "^1.0.0"
packaging = "^23.1"
parso = "^0.8.3"
pathspec = "^0.11.1"
platformdirs = "^3.8.0"
pyparsing = "^3.1.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
rdflib = "^6.3.2"
regex = "^2023.6.3"
requests = "^2.31.0"
sgmllib3k = "^1.0.0"
six = "^1.16.0"
text-unidecode = "^1.3"
tomli = "^2.0.1"
typing-extensions = "^4.7.0"
typish = "^1.9.3"
tzlocal = "^5.0.1"
unicode-slugify = "^0.1.5"
urllib3 = "^2.0.3"
webencodings = "^0.5.1"
zipp = "^3.15.0"
orgora = "^0.1.1"
cachetools-ext = "^0.0.8"
flask = "^2.3.2"
flask-cors = "^4.0.0"
flask-markdown = "^0.3"
flask-wtf = "^1.1.1"
jinja2 = "^3.1.2"
markdown = "^3.4.3"
markupsafe = "^2.1.3"
python-levenshtein = "^0.21.1"
pyyaml = "^6.0"
unidecode = "^1.3.6"
werkzeug = "^2.3.6"
wtforms = "^3.0.1"
sqlalchemy = "^2.0.23"
pytest = "^7.4.3"
gitpython = "^3.1.40"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
